test/e2e: Add image cleanup workaround
Due to an issue with containerd not caching images with snapshotter labels
and peer-pods using the nydus-snapshotter to implement the image pull on
guest, we see failures do to the pause image (and potentially other test
images) having been pulled with overlayfs. This means that nydus-snapshotter
doesn't trigger the guest_pull code, so to get this working, we first need
to clean up images that have already been cached locally. For the CI I've
written a debug container that does this, so run it after provisioning of
the cluster and hopefully the tests will then pass.

Signed-off-by: stevenhorsman <steven@uk.ibm.com>